Update to use JoinInfo type. Update usage of the next_diff functions. Update enclone_ranger version.

+++ enclone_tail/src/grouper.rs
@@ -80,7 +80,7 @@ pub fn grouper(
                 all.sort();
                 let mut i = 0;
                 while i < all.len() {
-                    let j = next_diff1_2(&all, i as i32) as usize;
+                    let j = next_diff1_2(&all, i);
                     let mut g = Vec::<usize>::new();
                     for k in i..j {
                         g.push(all[k].1);
@@ -121,7 +121,7 @@ pub fn grouper(
                 all.sort();
                 let mut i = 0;
                 while i < all.len() {
-                    let j = next_diff1_2(&all, i as i32) as usize;
+                    let j = next_diff1_2(&all, i);
                     let mut g = Vec::<usize>::new();
                     for k in i..j {
                         g.push(all[k].1);
@@ -152,7 +152,7 @@ pub fn grouper(
                 all.sort();
                 let mut i = 0;
                 while i < all.len() {
-                    let j = next_diff1_2(&all, i as i32) as usize;
+                    let j = next_diff1_2(&all, i);
                     if all[i].0.is_empty() {
                         for k in i..j {
                             let g = vec![all[k].1];
@@ -192,7 +192,7 @@ pub fn grouper(
                 all.sort();
                 let mut i = 0;
                 while i < all.len() {
-                    let j = next_diff1_2(&all, i as i32) as usize;
+                    let j = next_diff1_2(&all, i);
                     if all[i].0.is_empty() {
                         for k in i..j {
                             let g = vec![all[k].1];
@@ -240,7 +240,7 @@ pub fn grouper(
                 all.sort();
                 let mut i = 0;
                 while i < all.len() {
-                    let j = next_diff1_2(&all, i as i32) as usize;
+                    let j = next_diff1_2(&all, i);
                     if all[i].0.is_empty() {
                         for k in i..j {
                             let g = vec![all[k].1];
@@ -277,7 +277,7 @@ pub fn grouper(
                 all.sort();
                 let mut i = 0;
                 while i < all.len() {
-                    let j = next_diff1_2(&all, i as i32) as usize;
+                    let j = next_diff1_2(&all, i);
                     let mut g = Vec::<usize>::new();
                     for k in i..j {
                         g.push(all[k].1);
@@ -307,7 +307,7 @@ pub fn grouper(
                 all.sort();
                 let mut i = 0;
                 while i < all.len() {
-                    let j = next_diff1_2(&all, i as i32) as usize;
+                    let j = next_diff1_2(&all, i);
                     let mut g = Vec::<usize>::new();
                     for k in i..j {
                         g.push(all[k].1);
@@ -339,7 +339,7 @@ pub fn grouper(
                 all.sort();
                 let mut i = 0;
                 while i < all.len() {
-                    let j = next_diff1_2(&all, i as i32) as usize;
+                    let j = next_diff1_2(&all, i);
                     let mut g = Vec::<usize>::new();
                     for k in i..j {
                         g.push(all[k].1);
@@ -371,7 +371,7 @@ pub fn grouper(
                 all.sort();
                 let mut i = 0;
                 while i < all.len() {
-                    let j = next_diff1_2(&all, i as i32) as usize;
+                    let j = next_diff1_2(&all, i);
                     let mut g = Vec::<usize>::new();
                     for k in i..j {
                         g.push(all[k].1);
